--- a/ape_optimism/__init__.py
+++ b/ape_optimism/__init__.py
@@ -1,23 +1,31 @@
 from ape import plugins
-from ape.api.networks import LOCAL_NETWORK_NAME, ForkedNetworkAPI, NetworkAPI, create_network_type
-from ape_node import Node
-from ape_test import LocalProvider
-
-from .ecosystem import NETWORKS, Optimism, OptimismConfig
 
 
 @plugins.register(plugins.Config)
 def config_class():
+    from ape_optimism.ecosystem import OptimismConfig
+
     return OptimismConfig
 
 
 @plugins.register(plugins.EcosystemPlugin)
 def ecosystems():
+    from ape_optimism.ecosystem import Optimism
+
     yield Optimism
 
 
 @plugins.register(plugins.NetworkPlugin)
 def networks():
+    from ape.api.networks import (
+        LOCAL_NETWORK_NAME,
+        ForkedNetworkAPI,
+        NetworkAPI,
+        create_network_type,
+    )
+
+    from ape_optimism.ecosystem import NETWORKS
+
     for network_name, network_params in NETWORKS.items():
         yield "optimism", network_name, create_network_type(*network_params)
         yield "optimism", f"{network_name}-fork", ForkedNetworkAPI
@@ -28,6 +36,12 @@ def networks():
 
 @plugins.register(plugins.ProviderPlugin)
 def providers():
+    from ape.api.networks import LOCAL_NETWORK_NAME
+    from ape_node import Node
+    from ape_test import LocalProvider
+
+    from ape_optimism.ecosystem import NETWORKS
+
     for network_name in NETWORKS:
         yield "optimism", network_name, Node