### What I did

before:

```
In [1]: %time import ape_optimism
exitCPU times: user 1.75 s, sys: 175 ms, total: 1.92 s
Wall time: 1.92 s
```

after:

```
In [1]: %time import ape_optimism
CPU times: user 8.04 ms, sys: 3.53 ms, total: 11.6 ms
Wall time: 10.7 ms
```

### How I did it

### How to verify it

### Checklist

- [ ] Passes all linting checks (pre-commit and CI jobs)
- [ ] New test cases have been added and are passing
- [ ] Documentation has been updated
- [ ] PR title follows [Conventional Commit](https://www.conventionalcommits.org/en/v1.0.0/) standard (will be automatically included in the changelog)